requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
flask_cors==3.0.10
requests==2.30.0
flask_bcrypt
flask_jwt_extended
mysql-connector-python
